+++ Changelog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+ - FOR-2290: Ability to specify file name template for File component
+
 ## 3.20.17
 ### Fixed
  - Issue with the onChange for textarea's with wysiwyg enabled.